Various fixes from 64bit review
Replaced timing for profiler with the standard timer from platform_time
Replaced some int stuff with size_t

+++ loom/common/platform/platformIO.c
@@ -371,7 +371,7 @@ static int ftwWalker(const char *fpath, const struct stat *sb, int typeflag)
 static void platform_walkDirectory_r(const char *path, platform_subdirectoryWalkerCallback cb, void *payload)
 {
     char relativePath[MAX_PATH], searchPath[MAX_PATH];
-    int  l;
+    size_t l;
 
     // Time to walk directories. Stick \* on the end so it searches properly.
     WIN32_FIND_DATAA findData;
@@ -401,7 +401,7 @@ static void platform_walkDirectory_r(const char *path, platform_subdirectoryWalk
 
         // Great, a hit. Figure out the relative path.
         strcpy(relativePath, path);
-        l = (int)strlen(relativePath);
+        l = strlen(relativePath);
         relativePath[l]     = '/';
         relativePath[l + 1] = 0;
         strcpy(relativePath + l + 1, findData.cFileName);
@@ -436,7 +436,7 @@ void platform_walkSubdirectories(const char *rootPath, platform_subdirectoryWalk
 static void platform_walkFiles_r(const char *path, platform_fileWalkerCallback cb, void *payload)
 {
     char relativePath[MAX_PATH], searchPath[MAX_PATH];
-    int  l;
+    size_t l;
 
     // Time to walk directories. Stick \* on the end so it searches properly.
     WIN32_FIND_DATAA findData;
@@ -460,7 +460,7 @@ static void platform_walkFiles_r(const char *path, platform_fileWalkerCallback c
 
         // Great, a hit. Figure out the relative path.
         strcpy(relativePath, path);
-        l = (int)strlen(relativePath);
+        l = strlen(relativePath);
         relativePath[l]     = '/';
         relativePath[l + 1] = 0;
         strcpy(relativePath + l + 1, findData.cFileName);
